Use nokogiri to select the 1st child element
Since the article has the body parsed already, why not take advantage of that,
and use the awesome Nokogiri to select the first child element. This is a
flexible selector, which means it *could* get some code, or a flash object,
or some <script> element. It could be easily remedied.